Refactor: Moved 2 hooks inside hooks/Graph and edited the export on index.ts only
