+++ app/src/hooks/index.ts
@@ -14,8 +14,8 @@ export { useDocumentDescription } from './useDocumentDescription'
 export { useDocumentTitle } from './useDocumentTitle'
 export { useFundingBalance } from './useFundingBalance'
 export { useGraphMarketMakerData } from './Graph/useGraphMarketMakerData'
-export { useGraphMarketUserTxData } from './useGraphMarketUserTxData'
-export { useGraphMeta } from './useGraphMeta'
+export { useGraphMarketUserTxData } from './Graph/useGraphMarketUserTxData'
+export { useGraphMeta } from './Graph/useGraphMeta'
 export { useMarketMakerData } from './useMarketMakerData'
 export { usePolling } from './usePolling'
 export { useTokens } from './useTokens'
